+++ util/src/main/scala/LangTag.scala
@@ -4,11 +4,13 @@ import java.util.Locale
 
 /** Extractor for Locales, e.g. for use in pattern-matching request paths. */
 object LangTag {
-  def unapply(s: String): Option[Locale] = {
-    val locale = Locale.forLanguageTag(s)
+  class LocaleOpt(val locale: Locale) extends AnyVal {
     // if toLanguageTag returns "und", it means the language tag is undefined
-    if (locale.toLanguageTag().equals("und")) None else Some(locale)
+    def isEmpty: Boolean = "und" == locale.toLanguageTag
+    def get: Locale = locale
   }
+
+  def unapply(s: String): LocaleOpt = new LocaleOpt(Locale.forLanguageTag(s))
 
   def invalidLangTagMessage(invalidLangTag: String) = s"Invalid language tag: '$invalidLangTag'"
 }